Several updates
Added new method: UINavigateToNextView. This allows you to navigate to
any view from, say, a button. This method will add the current view to
the navigation stack so that the back button will take you back to this
point. This allows you to break out of the strict flow of navigation
lists to allow non-linear navigation through the views.
Added new example: navigation-list-3a.html. This illustrates the use of
UINavigateToNewView.
Added new method: resetApp. This resets the navigation history and
sends the user back to the main view. By passing a value of "hard" you
can for the browser to reload the app.
Updated the deletion list structure to name its delete and edit buttons
uniquely.
Fixed bug in deletion list where the scroll panel was not getting
refreshed properly.
Added new method: UIResetDeletionList. This allows you to reset the
deletion list at any time, such as if the use initiates a new task
while the deletion list is active.
Add new cell tag: cell value. This is for use like the date timee tag.
It gets styled and place like it but is for other values, such as
currency or units.